+++ site_scons/prereq_tools/base.py
@@ -595,7 +595,7 @@ class PreReqComponent():
 
     def run_build(self, opts):
         """Build and dependencies"""
-        common_reqs = ['ofi', 'hwloc', 'mercury', 'boost', 'uuid', 'crypto', 'protobufc',
+        common_reqs = ['ofi', 'hwloc', 'mercury', 'uuid', 'crypto', 'protobufc',
                        'lz4', 'isal', 'isal_crypto', 'argobots']
         client_reqs = ['fused', 'json-c', 'capstone', 'aio']
         server_reqs = ['pmdk', 'spdk', 'ipmctl']
