+++ src/mgmt/rpc.h
@@ -1,6 +1,6 @@
 /**
  * (C) Copyright 2016-2024 Intel Corporation.
- * (C) Copyright 2025 Hewlett Packard Enterprise Development LP
+ * (C) Copyright 2025-2026 Hewlett Packard Enterprise Development LP
  *
  * SPDX-License-Identifier: BSD-2-Clause-Patent
  */
@@ -13,6 +13,8 @@
 
 #include <daos/rpc.h>
 #include <daos/rsvc.h>
+
+#include <cart/rpc_gen_macros.h>
 
 /*
  * RPC operation codes
