+++ src/pipeline/pipeline_rpc.h
@@ -1,6 +1,6 @@
 /**
  * (C) Copyright 2016-2023 Intel Corporation.
- * (C) Copyright 2025 Hewlett Packard Enterprise Development LP
+ * (C) Copyright 2025-2026 Hewlett Packard Enterprise Development LP
  *
  * SPDX-License-Identifier: BSD-2-Clause-Patent
  */
@@ -11,6 +11,9 @@
 #include <daos/rpc.h>
 #include <daos/pipeline.h>
 #include <daos/object.h>
+
+#include <cart/rpc_gen_macros.h>
+
 #include "pipeline_internal.h"
 
 /* LIST of internal RPCS in form of:
