+++ src/rdb/tests/rpc.h
@@ -1,6 +1,6 @@
 /**
  * (C) Copyright 2017-2021 Intel Corporation.
- * (C) Copyright 2025 Hewlett Packard Enterprise Development LP
+ * (C) Copyright 2025-2026 Hewlett Packard Enterprise Development LP
  *
  * SPDX-License-Identifier: BSD-2-Clause-Patent
  */
@@ -9,6 +9,8 @@
 #define RDB_TESTS_RPC_H
 
 #include <daos/rsvc.h>
+
+#include <cart/rpc_gen_macros.h>
 
 /*
  * RPC operation codes
