+++ src/rsvc/rpc.h
@@ -1,6 +1,6 @@
 /*
  * (C) Copyright 2016-2022 Intel Corporation.
- * (C) Copyright 2025 Hewlett Packard Enterprise Development LP
+ * (C) Copyright 2025-2026 Hewlett Packard Enterprise Development LP
  *
  * SPDX-License-Identifier: BSD-2-Clause-Patent
  */
@@ -18,6 +18,8 @@
 #include <daos/rpc.h>
 #include <daos/rsvc.h>
 #include <daos_srv/rdb.h>
+
+#include <cart/rpc_gen_macros.h>
 
 /*
  * RPC operation codes
