+++ utils/rpms/daos.changelog
@@ -1,5 +1,8 @@
 %changelog
-* Tue Aug 11 2026  Jan Michalski <jan-marian.michalski@hpe.com@hpe.com> - 2.9.100-7
+* Mon Aug 17 2026 Jerome Soumagne <jerome.soumagne@hpe.com> 2.9.100-8
+- Remove boost-devel dependency (now uses internal headers)
+
+* Tue Aug 11 2026  Jan Michalski <jan-marian.michalski@hpe.com> - 2.9.100-7
 - Update PMDK to version 2.1.4
 
 * Fri Jul 10 2026 Kris Jacque <kris.jacque@hpe.com> 2.9.100-6
