+++ utils/rpms/daos.spec
@@ -25,7 +25,7 @@
 
 Name:          daos
 Version:       2.9.100
-Release:       7%{?relval}%{?dist}
+Release:       8%{?relval}%{?dist}
 Summary:       DAOS Storage Engine
 Obsoletes:     daos-firmware
 
@@ -56,11 +56,9 @@ BuildRequires: bullseye
 %if (0%{?rhel} >= 8)
 BuildRequires: argobots-devel >= %{argobots_version}
 BuildRequires: json-c-devel
-BuildRequires: boost-python3-devel
 %else
 BuildRequires: libabt-devel >= %{argobots_version}
 BuildRequires: libjson-c-devel
-BuildRequires: boost-devel
 %endif
 %if %{with server}
 BuildRequires: libpmemobj-devel >= 2.1.3
