+++ utils/rpms/mercury.changelog
@@ -1,4 +1,7 @@
 %changelog
+* Tue Sep 01 2026 Jerome Soumagne <jerome.soumagne@hpe.com> - 2.4.1-6
+- Remove boost external dependency and use provided boost headers instead
+
 * Tue Sep 01 2026 Jerome Soumagne <jerome.soumagne@hpe.com> - 2.4.1-5
 - Add patch to fix handling of canceled events
