+++ utils/scripts/install-el9.sh
@@ -18,7 +18,6 @@ dnf_install_args="${1:-}"
 
 # shellcheck disable=SC2086
 dnf --nodocs install ${dnf_install_args} \
-    boost-python3-devel \
     bzip2 \
     capstone-devel \
     cmake \
