+++ utils/scripts/install-leap15.sh
@@ -20,7 +20,6 @@ dnf_install_args="${1:-}"
 
 # shellcheck disable=SC2086
 dnf --nodocs install ${dnf_install_args} \
-    boost-devel \
     bzip2 \
     cmake \
     createrepo_c \
